move interval and loading messages to gatsby-ssr
